+++ packages/plugin-rsc/examples/basic/src/routes/root.tsx
@@ -46,6 +46,8 @@ import { TestHmrClientDep2 } from './hmr-client-dep2/client'
 import { TestHmrClientDep3 } from './hmr-client-dep3/server'
 import { TestChunk2 } from './chunk2/server'
 import { TestUseId } from './use-id/server'
+import { TestClientError } from './client-error/client'
+import { TestServerError } from './server-error/server'
 
 export function Root(props: { url: URL }) {
   return (
@@ -78,6 +80,8 @@ export function Root(props: { url: URL }) {
         <TestHmrSwitchClient />
         <TestTemporaryReference />
         <TestServerActionError />
+        <TestClientError />
+        <TestServerError url={props.url} />
         <TestReplayConsoleLogs url={props.url} />
         <TestSuspense url={props.url} />
         <TestActionFromClient />